refactor(desktop): extract ChatPane leaf components (Phase 2/5)
- skeletons.tsx
- AttachmentList.tsx (with AttachmentImageThumb)
- ImageAttachmentPreviewModal.tsx
- ArtifactBrowserModal.tsx (with OutputArtifactIcon)
- AssistantTurn/status.tsx (LiveStatusLine, Ellipsis, Cursor,
  TraceTimelineStepEntry, ExecutionTimelineThinkingEntry)

ArtifactBrowserModal export preserved on the public surface via
re-export from ChatPane.tsx.
