+++ desktop/src/components/panes/ChatPane.tsx
@@ -195,6 +195,35 @@ import {
   historyMessagesInDisplayOrder,
   turnInputIdsFromHistoryMessages,
 } from "./ChatPane/helpers";
+import {
+  HistoryRestoreSkeleton,
+  IntegrationErrorBanner,
+} from "./ChatPane/skeletons";
+import {
+  AttachmentList,
+  formatAttachmentSize,
+} from "./ChatPane/AttachmentList";
+import { ImageAttachmentPreviewModal } from "./ChatPane/ImageAttachmentPreviewModal";
+import {
+  ArtifactBrowserModal,
+  OutputArtifactIcon,
+  outputBrowserFilterForOutput,
+  outputChangeLabel,
+  outputKindLabel,
+  outputMetadataNumber,
+  outputMetadataString,
+  outputSecondaryLabel,
+  sortOutputs,
+  sortOutputsLatestFirst,
+  dedupeOutputsForDisplay,
+} from "./ChatPane/ArtifactBrowserModal";
+import {
+  ExecutionTimelineThinkingEntry,
+  LiveStatusEllipsis,
+  LiveStatusLine,
+  StreamingCursor,
+  TraceTimelineStepEntry,
+} from "./ChatPane/AssistantTurn/status";
 
 export type {
   ChatAssistantSegment,
@@ -206,6 +235,8 @@ export {
   inputIdFromMessageId,
   historyMessagesInDisplayOrder,
   turnInputIdsFromHistoryMessages,
+  ArtifactBrowserModal,
+  sortOutputs,
 };
 
 function sessionUserId(
@@ -787,225 +818,6 @@ export function chatMessagesFromSessionState(params: {
     );
 }
 
-function formatAttachmentSize(sizeBytes: number) {
-  if (!Number.isFinite(sizeBytes) || sizeBytes <= 0) {
-    return "";
-  }
-  if (sizeBytes >= 1024 * 1024) {
-    return `${(sizeBytes / (1024 * 1024)).toFixed(1)} MB`;
-  }
-  if (sizeBytes >= 1024) {
-    return `${Math.round(sizeBytes / 1024)} KB`;
-  }
-  return `${sizeBytes} B`;
-}
-
-function outputMetadataString(
-  output: WorkspaceOutputRecordPayload,
-  key: string,
-) {
-  const value = output.metadata[key];
-  return typeof value === "string" ? value.trim() : "";
-}
-
-function outputMetadataNumber(
-  output: WorkspaceOutputRecordPayload,
-  key: string,
-) {
-  const value = output.metadata[key];
-  return typeof value === "number" && Number.isFinite(value) ? value : null;
-}
-
-function outputBrowserFilterForOutput(
-  output: WorkspaceOutputRecordPayload,
-): ArtifactBrowserFilter {
-  if (
-    outputMetadataString(output, "origin_type") === "app" ||
-    output.module_id
-  ) {
-    return "apps";
-  }
-  const category = outputMetadataString(output, "category");
-  if (category === "image") {
-    return "images";
-  }
-  if (category === "code") {
-    return "code";
-  }
-  if (category === "link") {
-    return "links";
-  }
-  return "documents";
-}
-
-function outputKindLabel(output: WorkspaceOutputRecordPayload) {
-  // .dashboard files outrank generic origin/category labels — they're
-  // a first-class artifact type with their own renderer.
-  const ext = outputFileExtensionFromTitle(output);
-  if (ext === "dashboard") {
-    return "Dashboard";
-  }
-  if (
-    outputMetadataString(output, "origin_type") === "app" ||
-    output.module_id
-  ) {
-    const artifactType = outputMetadataString(output, "artifact_type");
-    if (artifactType) {
-      return artifactType.charAt(0).toUpperCase() + artifactType.slice(1);
-    }
-    return "Artifact";
-  }
-  const category = outputMetadataString(output, "category");
-  if (category === "image") {
-    return "Image";
-  }
-  if (category === "code") {
-    return "Code file";
-  }
-  if (category === "link") {
-    return "Link";
-  }
-  if (category === "spreadsheet") {
-    return "Spreadsheet";
-  }
-  if (category === "document") {
-    return "Document";
-  }
-  return output.output_type === "document" ? "Document" : "File";
-}
-
-// Lightweight extension lookup that doesn't depend on the metadata
-// envelope (which agent-authored files often skip). Reads the title
-// suffix only — used inside outputKindLabel before metadata parsing.
-function outputFileExtensionFromTitle(
-  output: WorkspaceOutputRecordPayload,
-): string {
-  const fromTitle = output.title?.trim() ?? "";
-  const dotIndex = fromTitle.lastIndexOf(".");
-  if (dotIndex > 0 && dotIndex < fromTitle.length - 1) {
-    return fromTitle.slice(dotIndex + 1).toLowerCase();
-  }
-  const path = outputMetadataString(output, "file_path") ?? output.file_path ?? "";
-  const pathDot = path.lastIndexOf(".");
-  if (pathDot > 0 && pathDot < path.length - 1) {
-    return path.slice(pathDot + 1).toLowerCase();
-  }
-  return "";
-}
-
-function outputChangeLabel(output: WorkspaceOutputRecordPayload) {
-  const changeType = outputMetadataString(output, "change_type");
-  if (changeType === "created") {
-    return "Created";
-  }
-  if (changeType === "modified") {
-    return "Updated";
-  }
-  return "";
-}
-
-function outputSecondaryLabel(output: WorkspaceOutputRecordPayload) {
-  const parts = [outputKindLabel(output)];
-  const sizeLabel = formatAttachmentSize(
-    outputMetadataNumber(output, "size_bytes") ?? 0,
-  );
-  if (sizeLabel) {
-    parts.push(sizeLabel);
-  }
-  const timeLabel = chatMessageTimeLabel(output.created_at);
-  if (timeLabel) {
-    parts.push(timeLabel);
-  }
-  return parts.join(" · ");
-}
-
-export function sortOutputs(outputs: WorkspaceOutputRecordPayload[]) {
-  return [...dedupeOutputsForDisplay(outputs)].sort((left, right) => {
-    const leftTime = Date.parse(left.created_at || "") || 0;
-    const rightTime = Date.parse(right.created_at || "") || 0;
-    if (leftTime !== rightTime) {
-      return leftTime - rightTime;
-    }
-    return left.title.localeCompare(right.title);
-  });
-}
-
-function sortOutputsLatestFirst(outputs: WorkspaceOutputRecordPayload[]) {
-  return [...dedupeOutputsForDisplay(outputs)].sort((left, right) => {
-    const leftTime = Date.parse(left.created_at || "") || 0;
-    const rightTime = Date.parse(right.created_at || "") || 0;
-    if (leftTime !== rightTime) {
-      return rightTime - leftTime;
-    }
-    return left.title.localeCompare(right.title);
-  });
-}
-
-function outputDisplayDedupeKey(output: WorkspaceOutputRecordPayload) {
-  const filePath = output.file_path?.trim() ?? "";
-  if (filePath) {
-    return `path:${filePath}`;
-  }
-  const artifactId = output.artifact_id?.trim() ?? "";
-  if (artifactId) {
-    return `artifact:${artifactId}`;
-  }
-  const title = output.title?.trim().toLowerCase() ?? "";
-  if (title) {
-    return `title:${title}`;
-  }
-  return `id:${output.id}`;
-}
-
-function outputDisplayPriority(output: WorkspaceOutputRecordPayload) {
-  let score = 0;
-  const originType = outputMetadataString(output, "origin_type");
-  if (originType === "forwarded_subagent") {
-    score += 40;
-  } else if (originType === "runtime_tool") {
-    score += 35;
-  } else if (originType === "app") {
-    score += 30;
-  }
-
-  if (outputMetadataString(output, "artifact_type") === "report") {
-    score += 20;
-  }
-  if (!/\.[A-Za-z0-9]+$/.test(output.title?.trim() ?? "")) {
-    score += 5;
-  }
-  return score;
-}
-
-function shouldPreferOutputForDisplay(
-  candidate: WorkspaceOutputRecordPayload,
-  current: WorkspaceOutputRecordPayload,
-) {
-  const candidatePriority = outputDisplayPriority(candidate);
-  const currentPriority = outputDisplayPriority(current);
-  if (candidatePriority !== currentPriority) {
-    return candidatePriority > currentPriority;
-  }
-  const candidateCreatedAt = Date.parse(candidate.created_at || "") || 0;
-  const currentCreatedAt = Date.parse(current.created_at || "") || 0;
-  if (candidateCreatedAt !== currentCreatedAt) {
-    return candidateCreatedAt > currentCreatedAt;
-  }
-  return candidate.title.localeCompare(current.title) < 0;
-}
-
-function dedupeOutputsForDisplay(outputs: WorkspaceOutputRecordPayload[]) {
-  const preferredByKey = new Map<string, WorkspaceOutputRecordPayload>();
-  for (const output of outputs) {
-    const key = outputDisplayDedupeKey(output);
-    const current = preferredByKey.get(key);
-    if (!current || shouldPreferOutputForDisplay(output, current)) {
-      preferredByKey.set(key, output);
-    }
-  }
-  return [...preferredByKey.values()];
-}
-
 function sortMemoryUpdateProposals(
   proposals: MemoryUpdateProposalRecordPayload[],
 ) {
@@ -1017,14 +829,6 @@ function sortMemoryUpdateProposals(
     }
     return left.title.localeCompare(right.title);
   });
-}
-
-function attachmentButtonLabel(attachment: {
-  name: string;
-  size_bytes: number;
-}) {
-  const sizeLabel = formatAttachmentSize(attachment.size_bytes);
-  return sizeLabel ? `${attachment.name} (${sizeLabel})` : attachment.name;
 }
 
 function attachmentUploadPayload(
@@ -1974,32 +1778,6 @@ function extractToolErrorText(payload: Record<string, unknown>): string {
   }
 
   return "";
-}
-
-function isIntegrationError(
-  text: string,
-): { provider: string; action: string } | null {
-  const patterns: Array<{ pattern: RegExp; provider: string }> = [
-    { pattern: /no\s+google\s+token/i, provider: "Google" },
-    { pattern: /no\s+github\s+token/i, provider: "GitHub" },
-    { pattern: /no\s+reddit\s+token/i, provider: "Reddit" },
-    { pattern: /no\s+twitter\s+token/i, provider: "Twitter" },
-    { pattern: /no\s+linkedin\s+token/i, provider: "LinkedIn" },
-    { pattern: /PLATFORM_INTEGRATION_TOKEN/i, provider: "" },
-    { pattern: /integration.*not.*connected/i, provider: "" },
-    { pattern: /integration.*not.*bound/i, provider: "" },
-    { pattern: /connect\s+via\s+(settings|integrations)/i, provider: "" },
-  ];
-  for (const { pattern, provider } of patterns) {
-    if (pattern.test(text)) {
-      const resolved = provider || "this provider";
-      return {
-        provider: resolved,
-        action: `Connect ${resolved} in the Integrations tab`,
-      };
-    }
-  }
-  return null;
 }
 
 function toolTraceStepFromPayload(
@@ -8766,257 +8544,6 @@ function AssistantTurnComponent({
   );
 }
 
-type OutputVisualKind =
-  | "spreadsheet"
-  | "document"
-  | "pdf"
-  | "code"
-  | "image"
-  | "link"
-  | "app"
-  | "dashboard"
-  | "file";
-
-const DASHBOARD_EXTENSIONS = new Set(["dashboard"]);
-
-const SPREADSHEET_EXTENSIONS = new Set([
-  "xlsx",
-  "xls",
-  "xlsm",
-  "xlsb",
-  "ods",
-  "csv",
-  "tsv",
-]);
-const PDF_EXTENSIONS = new Set(["pdf"]);
-const CODE_EXTENSIONS = new Set([
-  "ts",
-  "tsx",
-  "js",
-  "jsx",
-  "py",
-  "rb",
-  "go",
-  "rs",
-  "java",
-  "c",
-  "cpp",
-  "h",
-  "cs",
-  "php",
-  "swift",
-  "kt",
-  "sh",
-  "json",
-  "yml",
-  "yaml",
-  "toml",
-  "xml",
-  "sql",
-  "css",
-  "scss",
-]);
-const DOCUMENT_EXTENSIONS = new Set([
-  "md",
-  "mdx",
-  "markdown",
-  "txt",
-  "doc",
-  "docx",
-  "rtf",
-  "odt",
-  "html",
-  "htm",
-]);
-
-function outputFileExtension(output: WorkspaceOutputRecordPayload): string {
-  const metadataExt = outputMetadataString(output, "extension");
-  if (metadataExt) {
-    return metadataExt.replace(/^\./, "").toLowerCase();
-  }
-  const fromTitle = output.title?.trim() ?? "";
-  const dotIndex = fromTitle.lastIndexOf(".");
-  if (dotIndex > 0 && dotIndex < fromTitle.length - 1) {
-    return fromTitle.slice(dotIndex + 1).toLowerCase();
-  }
-  return "";
-}
-
-function outputVisualKind(
-  output: WorkspaceOutputRecordPayload,
-): OutputVisualKind {
-  const filter = outputBrowserFilterForOutput(output);
-  if (filter === "apps") {
-    return "app";
-  }
-  if (filter === "images") {
-    return "image";
-  }
-  if (filter === "links") {
-    return "link";
-  }
-
-  const extension = outputFileExtension(output);
-  if (extension) {
-    if (DASHBOARD_EXTENSIONS.has(extension)) {
-      return "dashboard";
-    }
-    if (SPREADSHEET_EXTENSIONS.has(extension)) {
-      return "spreadsheet";
-    }
-    if (PDF_EXTENSIONS.has(extension)) {
-      return "pdf";
-    }
-    if (CODE_EXTENSIONS.has(extension)) {
-      return "code";
-    }
-    if (DOCUMENT_EXTENSIONS.has(extension)) {
-      return "document";
-    }
-  }
-
-  if (filter === "code") {
-    return "code";
-  }
-  const category = outputMetadataString(output, "category");
-  if (category === "spreadsheet") {
-    return "spreadsheet";
-  }
-  if (category === "document") {
-    return "document";
-  }
-  return "file";
-}
-
-function outputVisualTheme(kind: OutputVisualKind): {
-  Icon: typeof FileText;
-  tileClass: string;
-  iconClass: string;
-} {
-  switch (kind) {
-    case "spreadsheet":
-      return {
-        Icon: FileSpreadsheet,
-        tileClass: "bg-success/12 ring-1 ring-inset ring-success/20",
-        iconClass: "text-success",
-      };
-    case "pdf":
-      return {
-        Icon: FileType,
-        tileClass: "bg-destructive/12 ring-1 ring-inset ring-destructive/20",
-        iconClass: "text-destructive",
-      };
-    case "document":
-      return {
-        Icon: FileText,
-        tileClass: "bg-info/12 ring-1 ring-inset ring-info/20",
-        iconClass: "text-info",
-      };
-    case "code":
-      return {
-        Icon: FileCode2,
-        tileClass: "bg-info/12 ring-1 ring-inset ring-info/20",
-        iconClass: "text-info",
-      };
-    case "image":
-      return {
-        Icon: ImageIcon,
-        tileClass: "bg-warning/12 ring-1 ring-inset ring-warning/20",
-        iconClass: "text-warning",
-      };
-    case "link":
-      return {
-        Icon: Link2,
-        tileClass: "bg-info/12 ring-1 ring-inset ring-info/20",
-        iconClass: "text-info",
-      };
-    case "app":
-      return {
-        Icon: Waypoints,
-        tileClass: "bg-primary/12 ring-1 ring-inset ring-primary/20",
-        iconClass: "text-primary",
-      };
-    case "dashboard":
-      return {
-        Icon: LayoutDashboard,
-        tileClass: "bg-primary/12 ring-1 ring-inset ring-primary/20",
-        iconClass: "text-primary",
-      };
-    default:
-      return {
-        Icon: FileIcon,
-        tileClass: "bg-muted ring-1 ring-inset ring-border",
-        iconClass: "text-muted-foreground",
-      };
-  }
-}
-
-function OutputArtifactIcon({
-  output,
-  size = "md",
-}: {
-  output: WorkspaceOutputRecordPayload;
-  size?: "sm" | "md";
-}) {
-  const kind = outputVisualKind(output);
-  const { Icon, tileClass, iconClass } = outputVisualTheme(kind);
-  const tileSize = size === "sm" ? "size-7" : "size-9";
-  const iconSize = size === "sm" ? 14 : 16;
-  return (
-    <div
-      className={`grid ${tileSize} shrink-0 place-items-center rounded-lg ${tileClass}`}
-    >
-      <Icon size={iconSize} className={iconClass} />
-    </div>
-  );
-}
-
-function HistoryRestoreSkeleton() {
-  return (
-    <div
-      role="status"
-      aria-label="Loading conversation"
-      className="absolute inset-0 z-30 overflow-hidden px-6 pb-5 pt-5"
-    >
-      <div className="flex h-full flex-col">
-        <div className="animate-pulse space-y-6">
-          <div className="flex items-start justify-between gap-6">
-            <div className="h-5 w-28 rounded-md bg-muted" />
-            <div className="h-11 w-52 rounded-2xl bg-muted" />
-          </div>
-          <div className="space-y-3 px-3">
-            <div className="flex items-center gap-2">
-              <div className="h-5 w-6 rounded-md bg-muted" />
-              <div className="h-5 w-14 rounded-md bg-muted" />
-            </div>
-            <div className="h-5 w-full rounded-md bg-muted" />
-            <div className="h-5 w-full rounded-md bg-muted" />
-            <div className="h-5 w-[42%] rounded-md bg-muted" />
-          </div>
-        </div>
-
-        <div className="mt-auto">
-          <div className="rounded-2xl border border-border bg-muted p-4">
-            <div className="animate-pulse space-y-3">
-              <div className="h-6 w-full rounded-lg bg-muted" />
-              <div className="flex items-center justify-between gap-3">
-                <div className="flex items-center gap-2">
-                  <div className="size-8 rounded-full bg-muted" />
-                  <div className="size-8 rounded-full bg-muted" />
-                </div>
-                <div className="flex items-center gap-2">
-                  <div className="size-8 rounded-full bg-muted" />
-                  <div className="size-8 rounded-full bg-muted" />
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
 function AssistantTurnOutputs({
   outputs,
   onOpenOutput,
@@ -9368,150 +8895,6 @@ export function ConversationTurns<Message extends ChatMessage>({
   );
 }
 
-export function ArtifactBrowserModal({
-  open,
-  filter,
-  outputs,
-  scope,
-  onClose,
-  onFilterChange,
-  onOpenOutput,
-  layout = "page",
-}: {
-  open: boolean;
-  filter: ArtifactBrowserFilter;
-  outputs: WorkspaceOutputRecordPayload[];
-  scope: "session" | "reply";
-  onClose: () => void;
-  onFilterChange: (nextFilter: ArtifactBrowserFilter) => void;
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  layout?: "page" | "card";
-}) {
-  if (!open) {
-    return null;
-  }
-
-  const filterLabels: Array<{ id: ArtifactBrowserFilter; label: string }> = [
-    { id: "all", label: "All" },
-    { id: "documents", label: "Documents" },
-    { id: "images", label: "Images" },
-    { id: "code", label: "Code files" },
-    { id: "links", label: "Links" },
-    { id: "apps", label: "Apps" },
-  ];
-  const allDisplayOutputs =
-    outputs.length > 1 ? dedupeOutputsForDisplay(outputs) : outputs;
-  const filteredOutputs = sortOutputsLatestFirst(
-    filter === "all"
-      ? allDisplayOutputs
-      : allDisplayOutputs.filter(
-          (output) => outputBrowserFilterForOutput(output) === filter,
-        ),
-  );
-  const overlayClassName =
-    layout === "card"
-      ? "absolute inset-0 z-30 flex items-stretch justify-stretch bg-background/88 p-2 backdrop-blur-[2px]"
-      : "absolute inset-0 z-30 flex items-center justify-center bg-black/40 px-6 py-8 backdrop-blur-[2px]";
-  const panelClassName =
-    layout === "card"
-      ? "flex h-full w-full min-h-0 flex-col overflow-hidden rounded-2xl border border-border bg-background shadow-xl"
-      : "flex max-h-full w-full max-w-lg flex-col overflow-hidden rounded-xl border border-border bg-background shadow-xl";
-
-  return (
-    <div
-      className={overlayClassName}
-      data-control-center-swipe-ignore={layout === "card" ? "true" : undefined}
-    >
-      <div className={panelClassName}>
-        <div className="flex shrink-0 items-center justify-between gap-3 border-b border-border px-4 py-3">
-          <div>
-            <div className="text-sm font-semibold text-foreground">
-              Artifacts
-            </div>
-            <div className="text-xs text-muted-foreground">
-              {allDisplayOutputs.length} item
-              {allDisplayOutputs.length === 1 ? "" : "s"}{" "}
-              {scope === "reply"
-                ? "attached to this reply"
-                : "in this session"}
-            </div>
-          </div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={onClose}
-            aria-label="Close"
-          >
-            <X className="size-3.5" />
-          </Button>
-        </div>
-
-        <div className="flex shrink-0 flex-wrap gap-1 border-b border-border px-3 py-2">
-          {filterLabels.map((item) => {
-            const active = filter === item.id;
-            return (
-              <Button
-                key={item.id}
-                variant={active ? "secondary" : "ghost"}
-                size="xs"
-                onClick={() => onFilterChange(item.id)}
-              >
-                {item.label}
-              </Button>
-            );
-          })}
-        </div>
-
-        <div className="min-h-0 flex-1 overflow-y-auto px-3 py-3">
-          {filteredOutputs.length === 0 ? (
-            <div className="py-10 text-center text-xs text-muted-foreground">
-              No artifacts match this filter.
-            </div>
-          ) : (
-            <div className="grid gap-2">
-              {filteredOutputs.map((output) => {
-                const kindLabel = outputKindLabel(output);
-                const changeLabel = outputChangeLabel(output);
-                return (
-                  <button
-                    key={output.id}
-                    type="button"
-                    onClick={() => {
-                      onClose();
-                      onOpenOutput?.(output);
-                    }}
-                    disabled={!onOpenOutput}
-                    className="group flex w-full min-w-0 items-start gap-2.5 rounded-lg border border-border bg-card px-3 py-2.5 text-left ring-border transition-colors hover:bg-accent/50 disabled:cursor-default disabled:hover:bg-card"
-                  >
-                    <OutputArtifactIcon output={output} />
-                    <div className="min-w-0 flex-1">
-                      <div className="text-[10px] font-medium uppercase tracking-wide text-muted-foreground">
-                        {kindLabel}
-                      </div>
-                      <div className="truncate text-sm font-medium text-foreground">
-                        {output.title || "Untitled artifact"}
-                      </div>
-                    </div>
-                    {changeLabel ? (
-                      <Badge
-                        variant="outline"
-                        className="shrink-0 text-[10px] uppercase"
-                      >
-                        {changeLabel}
-                      </Badge>
-                    ) : null}
-                  </button>
-                );
-              })}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
 function traceStatusLabel(status: ChatTraceStepStatus) {
   if (status === "completed") {
     return "Completed";
@@ -9525,78 +8908,6 @@ function traceStatusLabel(status: ChatTraceStepStatus) {
   return "In progress";
 }
 
-function isTraceStepCollapsed(
-  step: ChatTraceStep,
-  collapsedTraceByStepId: Record<string, boolean>,
-) {
-  return collapsedTraceByStepId[step.id] ?? true;
-}
-
-function IntegrationErrorBanner({ details }: { details: string[] }) {
-  const errorText = details.join(" ");
-  const integrationError = isIntegrationError(errorText);
-  if (!integrationError) return null;
-  return (
-    <div className="mt-1.5 flex items-center gap-2 rounded-lg border border-warning/20 bg-warning/8 px-2.5 py-1.5 text-xs text-warning">
-      <Cable className="size-3 shrink-0" />
-      <span>{integrationError.action}</span>
-    </div>
-  );
-}
-
-function LiveStatusEllipsis() {
-  return (
-    <span
-      aria-hidden="true"
-      className="inline-flex shrink-0 items-center text-muted-foreground"
-    >
-      <DotmSquare3 dotSize={1} size={10} />
-    </span>
-  );
-}
-
-function StreamingCursor() {
-  return (
-    <>
-      <style>{`
-        @keyframes streaming-cursor-blink {
-          0%, 50% { opacity: 1; }
-          50.01%, 100% { opacity: 0; }
-        }
-      `}</style>
-      <span
-        aria-hidden="true"
-        className="ml-0.5 inline-block h-[1em] w-[2px] -mb-[2px] translate-y-[3px] rounded-[1px] bg-foreground/65"
-        style={{ animation: "streaming-cursor-blink 1100ms steps(1) infinite" }}
-      />
-    </>
-  );
-}
-
-function LiveStatusLine({
-  label,
-  className = "",
-}: {
-  label: string;
-  className?: string;
-}) {
-  const normalizedLabel = label.replace(/\.+$/, "").trim();
-  if (!normalizedLabel) {
-    return null;
-  }
-
-  return (
-    <div
-      aria-live="polite"
-      key={normalizedLabel}
-      className={`flex w-fit items-center gap-1.5 text-xs leading-none text-muted-foreground animate-in fade-in-0 slide-in-from-bottom-0.5 duration-200 ease-out ${className}`.trim()}
-    >
-      <LiveStatusEllipsis />
-      <span>{normalizedLabel}</span>
-    </div>
-  );
-}
-
 function summarizeThinking(text: string) {
   const firstContentLine =
     text
@@ -9607,85 +8918,6 @@ function summarizeThinking(text: string) {
   return firstContentLine.length > 88
     ? `${firstContentLine.slice(0, 85).trimEnd()}...`
     : firstContentLine;
-}
-
-function TraceTimelineStepEntry({
-  step,
-  collapsedByStepId,
-  onToggleStep,
-}: {
-  step: ChatTraceStep;
-  collapsedByStepId: Record<string, boolean>;
-  onToggleStep: (stepId: string) => void;
-}) {
-  const expanded = !(collapsedByStepId[step.id] ?? true);
-
-  return (
-    <div>
-      <button
-        type="button"
-        onClick={() => step.details.length > 0 && onToggleStep(step.id)}
-        className={`flex w-full items-start gap-2 rounded-md px-2.5 -ml-2.5 py-1 text-left text-xs transition-colors ${step.details.length > 0 ? "hover:bg-muted cursor-pointer" : "cursor-default"}`}
-      >
-        <span className="mt-0.5 shrink-0">
-          {step.status === "completed" ? (
-            <Check className="size-3 text-success" />
-          ) : step.status === "error" ? (
-            <AlertTriangle className="size-3 text-destructive" />
-          ) : step.status === "running" ? (
-            <Loader2 className="size-3 animate-spin text-muted-foreground" />
-          ) : (
-            <Clock3 className="size-3 text-muted-foreground" />
-          )}
-        </span>
-        <span className="min-w-0 flex-1">
-          <span className="font-medium text-foreground">{step.title}</span>
-          {step.details.length > 0 ? (
-            <span className="ml-1.5 text-muted-foreground">
-              {step.details[0]}
-            </span>
-          ) : null}
-        </span>
-        {step.details.length > 1 ? (
-          <ChevronDown
-            className={`size-3 mt-0.5 shrink-0 text-muted-foreground transition-transform ${expanded ? "rotate-180" : ""}`}
-          />
-        ) : null}
-      </button>
-      {expanded && step.details.length > 1 ? (
-        <div className="ml-6 mt-0.5 mb-1 rounded-md border border-border bg-muted px-3 py-2 text-xs leading-5 text-muted-foreground whitespace-pre-wrap">
-          {step.details.slice(1).join("\n")}
-        </div>
-      ) : null}
-      {step.status === "error" ? (
-        <IntegrationErrorBanner details={step.details} />
-      ) : null}
-    </div>
-  );
-}
-
-function ExecutionTimelineThinkingEntry({
-  text,
-  onLinkClick,
-  onLocalLinkClick,
-}: {
-  text: string;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-}) {
-  return (
-    <div className="py-1">
-      <div className="-ml-2.5 w-[calc(100%+0.625rem)] rounded-xl border border-border bg-muted px-3.5 py-3">
-        <SimpleMarkdown
-          className="chat-markdown chat-thinking-markdown max-w-full text-foreground"
-          onLinkClick={onLinkClick}
-          onLocalLinkClick={onLocalLinkClick}
-        >
-          {text}
-        </SimpleMarkdown>
-      </div>
-    </div>
-  );
 }
 
 function TraceStepGroup({
@@ -9839,240 +9071,6 @@ function TraceStepGroup({
       ) : null}
     </div>
   );
-}
-
-function AttachmentImageThumb({ file }: { file: File }) {
-  const [objectUrl, setObjectUrl] = useState<string | null>(null);
-  const [loadFailed, setLoadFailed] = useState(false);
-
-  useEffect(() => {
-    const url = URL.createObjectURL(file);
-    setObjectUrl(url);
-    return () => URL.revokeObjectURL(url);
-  }, [file]);
-
-  if (loadFailed || !objectUrl) {
-    return <ImageIcon className="size-4 shrink-0 text-primary" />;
-  }
-  return (
-    <img
-      alt=""
-      className="size-7 shrink-0 rounded-md object-cover"
-      onError={() => setLoadFailed(true)}
-      src={objectUrl}
-    />
-  );
-}
-
-function AttachmentList({
-  attachments,
-  onRemove,
-  onPreview,
-  className = "",
-}: {
-  attachments: AttachmentListItem[];
-  onRemove?: (attachmentId: string) => void;
-  onPreview?: (attachment: AttachmentListItem) => void;
-  className?: string;
-}) {
-  return (
-    <div className={`flex flex-wrap gap-2 ${className}`.trim()}>
-      {attachments.map((attachment) => {
-        const isImagePreviewable =
-          attachment.kind === "image" &&
-          Boolean(onPreview) &&
-          Boolean(
-            attachment.file ||
-            (typeof attachment.workspace_path === "string" &&
-              attachment.workspace_path.trim()),
-          );
-
-        const icon =
-          attachment.kind === "image" && attachment.file ? (
-            <AttachmentImageThumb file={attachment.file} />
-          ) : attachment.kind === "image" ? (
-            <ImageIcon className="size-4 shrink-0 text-primary" />
-          ) : attachment.kind === "folder" ? (
-            <Folder className="size-3.5 shrink-0 text-primary" />
-          ) : (
-            <FileText className="size-3.5 shrink-0 text-primary" />
-          );
-
-        const labelClassName = "truncate";
-        const isImageThumb = attachment.kind === "image" && Boolean(attachment.file);
-
-        const content = (
-          <>
-            {icon}
-            <span className={labelClassName}>
-              {attachmentButtonLabel(attachment)}
-            </span>
-          </>
-        );
-
-        return (
-          <div
-            className="group/attachment bg-muted relative inline-flex max-w-full items-center gap-2 rounded-lg border border-border pr-2 text-xs text-foreground"
-            key={attachment.id}
-            style={{
-              paddingLeft: isImageThumb ? "4px" : "10px",
-              paddingTop: isImageThumb ? "4px" : "5px",
-              paddingBottom: isImageThumb ? "4px" : "5px",
-            }}
-          >
-            {isImagePreviewable ? (
-              <button
-                aria-label={`Preview ${attachment.name}`}
-                className="flex min-w-0 items-center gap-2 rounded-md text-left transition hover:text-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary/40"
-                onClick={() => onPreview?.(attachment)}
-                title={`Preview ${attachment.name}`}
-                type="button"
-              >
-                {content}
-              </button>
-            ) : (
-              content
-            )}
-            {onRemove ? (
-              <button
-                aria-label={`Remove ${attachment.name}`}
-                className="grid size-4 shrink-0 place-items-center rounded-full text-muted-foreground opacity-0 transition group-hover/attachment:opacity-100 hover:text-foreground"
-                onClick={() => onRemove(attachment.id)}
-                type="button"
-              >
-                <X className="size-3" />
-              </button>
-            ) : null}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-function ImageAttachmentPreviewModal({
-  open,
-  preview,
-  onClose,
-}: {
-  open: boolean;
-  preview: ImageAttachmentPreviewState | null;
-  onClose: () => void;
-}) {
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-
-    const handleKeyDown = (event: globalThis.KeyboardEvent) => {
-      if (event.key !== "Escape") {
-        return;
-      }
-      event.preventDefault();
-      onClose();
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [open, onClose]);
-
-  if (!open || !preview) {
-    return null;
-  }
-
-  const sizeLabel = formatAttachmentSize(preview.attachment.size_bytes);
-  const showImage = !preview.isLoading && !preview.errorMessage;
-  const modalContent = (
-    <div
-      className="fixed inset-0 z-[90] flex items-center justify-center px-6 py-8"
-      onClick={onClose}
-    >
-      {preview.browserSnapshot ? (
-        <img
-          aria-hidden="true"
-          src={preview.browserSnapshot.dataUrl}
-          alt=""
-          className="pointer-events-none absolute object-fill"
-          style={{
-            left: `${preview.browserSnapshot.bounds.x}px`,
-            top: `${preview.browserSnapshot.bounds.y}px`,
-            width: `${preview.browserSnapshot.bounds.width}px`,
-            height: `${preview.browserSnapshot.bounds.height}px`,
-          }}
-        />
-      ) : null}
-      <div
-        aria-hidden="true"
-        className="absolute inset-0 bg-black/70 backdrop-blur-[2px]"
-      />
-      <div
-        role="dialog"
-        aria-modal="true"
-        aria-label={`Preview ${preview.attachment.name}`}
-        className="relative z-10 flex max-h-[calc(100vh-64px)] flex-col overflow-hidden rounded-2xl border border-white/10 bg-background shadow-2xl"
-        style={{ maxWidth: "92vw" }}
-        onClick={(event) => event.stopPropagation()}
-      >
-        <div className="flex shrink-0 items-center justify-between gap-3 border-b border-border px-4 py-3">
-          <div className="min-w-0">
-            <div className="truncate text-sm font-semibold text-foreground">
-              {preview.attachment.name}
-            </div>
-            <div className="truncate text-xs text-muted-foreground">
-              {sizeLabel || "Image attachment"}
-            </div>
-          </div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={onClose}
-            aria-label="Close image preview"
-          >
-            <X className="size-3.5" />
-          </Button>
-        </div>
-
-        <div
-          className={`overflow-auto px-4 py-4 ${
-            showImage
-              ? "bg-transparent"
-              : "min-h-[240px] min-w-[320px] bg-muted/20"
-          }`}
-        >
-          {preview.isLoading ? (
-            <div className="flex h-full min-h-[208px] items-center justify-center gap-2 text-sm text-foreground/80">
-              <Loader2 className="size-4 animate-spin" />
-              <span>Loading preview...</span>
-            </div>
-          ) : preview.errorMessage ? (
-            <div className="flex h-full min-h-[208px] flex-col items-center justify-center gap-3 px-6 text-center">
-              <AlertTriangle className="size-5 text-warning" />
-              <p className="max-w-md text-sm text-foreground/80">
-                {preview.errorMessage}
-              </p>
-            </div>
-          ) : (
-            <div className="flex items-center justify-center">
-              <img
-                src={preview.dataUrl}
-                alt={preview.attachment.name}
-                className="block h-auto w-auto rounded-lg ring-1 ring-black/8"
-                style={{
-                  maxWidth: "calc(92vw - 32px)",
-                  maxHeight: "calc(88vh - 128px)",
-                }}
-              />
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-
-  return createPortal(modalContent, document.body);
 }
 
 function ModelCombobox({
